should bring this back inline with 1.5.8

+++ test/wsdl/overload/test_overload.rb
@@ -10,21 +10,21 @@ module WSDL; module Overload
 
 
 class TestOverload < Test::Unit::TestCase
-  TNS = "urn:overload"
+  TNS = "http://confluence.atlassian.com/rpc/soap-axis/confluenceservice-v1"
 
   Methods = [
     [
       XSD::QName.new(TNS, 'methodAlpha'), "methodAlpha1", "method_alpha_1",
-      [ [:in, "in0", ["::SOAP::SOAPString"]],
-        [:in, "in1", ["::SOAP::SOAPString"]],
-        [:in, "in2", ["::SOAP::SOAPString"]],
-        [:retval, "methodAlphaReturn", ["::SOAP::SOAPLong"]] ]
+      [ ["in", "in0", ["::SOAP::SOAPString"]],
+        ["in", "in1", ["::SOAP::SOAPString"]],
+        ["in", "in2", ["::SOAP::SOAPString"]],
+        ["retval", "methodAlphaReturn", ["::SOAP::SOAPLong"]] ]
     ], 
     [
       XSD::QName.new(TNS, 'methodAlpha'), "methodAlpha2", "method_alpha_2",
-      [ [:in, "in0", ["::SOAP::SOAPString"]],
-        [:in, "in1", ["::SOAP::SOAPString"]],
-        [:retval, "methodAlphaReturn", ["::SOAP::SOAPLong"]] ]
+      [ ["in", "in0", ["::SOAP::SOAPString"]],
+        ["in", "in1", ["::SOAP::SOAPString"]],
+        ["retval", "methodAlphaReturn", ["::SOAP::SOAPLong"]] ]
     ]
   ]
 
@@ -56,13 +56,7 @@ class TestOverload < Test::Unit::TestCase
 
   def teardown
     teardown_server if @server
-    unless $DEBUG
-      File.unlink(pathname('default.rb'))
-      File.unlink(pathname('defaultMappingRegistry.rb'))
-      File.unlink(pathname('defaultDriver.rb'))
-      File.unlink(pathname('defaultServant.rb'))
-      File.unlink(pathname('OverloadServiceClient.rb'))
-    end
+    File.unlink(pathname('default.rb')) unless $DEBUG
     @client.reset_stream if @client
   end
 
@@ -78,10 +72,6 @@ class TestOverload < Test::Unit::TestCase
     gen.basedir = DIR
     gen.logger.level = Logger::FATAL
     gen.opt['classdef'] = nil
-    gen.opt['mapping_registry'] = nil
-    gen.opt['driver'] = nil
-    gen.opt['servant_skelton'] = nil
-    gen.opt['client_skelton'] = nil
     gen.opt['force'] = true
     gen.run
     TestUtil.require(DIR, 'default.rb')
@@ -95,12 +85,6 @@ class TestOverload < Test::Unit::TestCase
 
   def pathname(filename)
     File.join(DIR, filename)
-  end
-
-  def test_compare
-    compare("expectedDriver.rb", "defaultDriver.rb")
-    compare("expectedServant.rb", "defaultServant.rb")
-    compare("expectedClient.rb", "OverloadServiceClient.rb")
   end
 
   def test_wsdl
@@ -120,10 +104,6 @@ class TestOverload < Test::Unit::TestCase
     @client.wiredump_dev = STDOUT if $DEBUG
     assert_equal(3, @client.call("methodAlpha1", "1", "2", "3"))
     assert_equal(2, @client.call("methodAlpha2", "1", "2"))
-  end
-
-  def compare(expected, actual)
-    TestUtil.filecompare(pathname(expected), pathname(actual))
   end
 end
